Add parsing of Universal Tables
Adds a symbols dictionary so that addresses can be converted to names.

Adds utility functions to create entries or sections that are offsets.
The value of such entries appear like this:
"-> 0x1234 (LABELNAME)"

Parsing of Universal Tables happens only if a "UNIVTABLES" entry exists
in the ROM map.

Note that not all fields of Universal Tables exist in all ROM versions.
Some care is taken so that the template does not do unreasonable things
with unexpected offsets and counts.