Ensure each stage has section depth 0
Sometimes, an error will cause the ending of a section to be skipped.
That would cause subsequent sections to be incorrectly placed inside the
wrong section. This change ensures that all the sections that are meant
to be in the root section are actually in the root section. This
includes DeclROM, Extended DeclROM, and System ROM.